feat(client):common catalog component for project and workspace
