+++ packages/amplication-client/src/Catalog/CatalogGrid.scss
@@ -1,6 +1,6 @@
 @import "../style/index.scss";
 
-.catalog {
+.catalog-grid {
   &__wrapper {
     display: flex;
     flex-direction: column;
